+++ osutil/disks/disks_darwin.go
@@ -32,6 +32,11 @@ var diskFromDeviceName = func(deviceName string) (Disk, error) {
 	return nil, osutil.ErrDarwin
 }
 
+// DiskFromPartitionDeviceNode is not implemented on darwin
+func DiskFromPartitionDeviceNode(node string) (Disk, error) {
+	return nil, osutil.ErrDarwin
+}
+
 // DiskFromDevicePath is not implemented on darwin
 func DiskFromDevicePath(devicePath string) (Disk, error) {
 	return nil, osutil.ErrDarwin
@@ -43,6 +48,10 @@ var diskFromDevicePath = func(devicePath string) (Disk, error) {
 
 // DiskFromMountPoint is not implemented on darwin
 func DiskFromMountPoint(mountpoint string, opts *Options) (Disk, error) {
+	return nil, osutil.ErrDarwin
+}
+
+func AllPhysicalDisks() ([]Disk, error) {
 	return nil, osutil.ErrDarwin
 }
